requirements.txt
beautifulsoup4==4.12.2
html5lib==1.1
numpy==1.26.1
pandas==2.1.1
Pillow==10.1.0
requests==2.31.0
torch==2.1.0
tqdm==4.66.1
triton==2.1.0
